requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
djlint==1.34.0
docopt-ng==0.9.0
EditorConfig==0.12.3
Flask==3.0.0
Flask-WTF==1.2.1
html-tag-names==0.1.2
html-void-elements==0.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.11
json5==0.9.14
MarkupSafe==2.1.3
pathspec==0.11.2
python-dotenv==1.0.0
python-http-client==3.3.7
PyYAML==6.0.1
regex==2023.10.3
sendgrid==6.10.0
six==1.16.0
starkbank-ecdsa==2.2.0
staticjinja==5.0.0
tomli==2.0.1
tqdm==4.66.1
watchdog==3.0.0
Werkzeug==3.0.1
WTForms==3.1.1
